Move sample caching code to own file fluid_samplecache.c/h
